Port telepathy-sofiasip to telepathy-glib 0.5.6 handle API.
This patch stops associating Telepathy handles with NUA handles using the
"hmagic" mechanism, because I can't see how to tell when SofiaSIP has finished
with a particular NUA handle (and hence how to not leak references to the
Telepathy handles). Instead I find out the Telepathy handle for text messages
or incoming events using the SIP headers, and find out whether a media channel
is the one we want by using the NUA handle as a unique key.

20070323174104-53eee-8906a0fe3be387b7d5e9a9f59b9cacae9b2696b4.gz