Always update the sending state in sip_media_stream_set_direction(),
even if the direction and pending flags properties did not change.
The reason is, we temporarily disable sending in sip_media_stream_set_remote_media() and call
sip_media_stream_set_direction() at the end to finalize the properties and the sending state.
In case the direction does not change, the sending state was left disabled.

20070831151942-5b6ca-4d8b5daa30d8cc6599edd76d8094df05676e82a7.gz