Pretend the call state is Ringing when we get 183 Session Progress
This is a quick fix to a bug where the remote media gateway starts sending
early audio with a ringback tone, but we're not getting any because of
a NAT or a firewall, which the local stream cannot puncture yet because
it doesn't yet send any media due to privacy reasons. A semi-good thing
to do in the stream engine would be to send some allowed, but harmless
packets such as comfort noise, but that requires quite a bit of
implementation.

The proper fix in the connection manager will wait for introduction
of a separate call state flag in the Telepathy specification and
telepathy-glib. See fd.o bug #23409.