Ask for a password in priv_handle_auth() if not given.
This requires a call to tp_simple_password_manager_prompt_async(), so
we create a data struct to help us split the function in half.

Also, we move all checks from the second half of the function into the
first, so they can be handled synchronously.

The diff is clearer if you use git diff --patience.