Fix group flags for streamedmedia calls.
We start off with lots of flags, and then delete them as we work out what
kind of channel we are, rather than trying to track what we need to
add/remove over time. We should always have the right flags before we are
advertised on the bus.

Note that adding members to a channel made using Create isn't *strictly*
correct way to do it, but it isn't special-cased in any of the code, and
I don't want to set flags that are lies so I've just updated the tests
accordingly.

Also, converting the flags to their binary representation makes failed
assertions a lot easier to debug (and I honestly don't care about speed).