configure.ac: make external package checks considerably simpler
PKG_CHECK_MODULES without a third and fourth argument already issues
an error if the package is missing, and its error messages are better
than ours. Also, we don't need to AC_SUBST the CFLAGS or LIBS, the
macro already does that for us. Also, stop ignoring the error (!)
if dbus-1 or dbus-glib-1 is missing. Finally, stop underquoting.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=49725
Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Reviewed-by: Mikhail Zabaluev <mikhail.zabaluev@nokia.com>